Fix nitpicks and refactor code to use antd components
